+++ tests/test_cli_manager.py
@@ -6,7 +6,7 @@ from tests.testcase import ManagedTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.7"
+__version__ = "0.1.0"
 
 
 class TestCliManager(ManagedTestCase):