+++ hookee/plugins/response_echo.py
@@ -1,10 +1,12 @@
+import json
+
 from flask import current_app
 
 from hookee import pluginmanager
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 
 plugin_type = pluginmanager.RESPONSE_PLUGIN
 print_util = None
@@ -18,9 +20,17 @@ def setup(cli_manager):
 
 def run(request, response):
     if not response:
+        data = ""
+        content_type = request.headers.get("Content-Type")
+        if request.form and not request.data:
+            data = json.dumps(dict(request.form))
+            content_type = "application/json"
+        elif request.data:
+            data = request.data.decode("utf-8")
+
         response = current_app.response_class(
-            "<?xml version=\"1.0\" encoding=\"UTF-8\"?><Response>Ok</Response>",
-            mimetype="application/xml",
+            data,
+            mimetype=content_type,
         )
 
     return response