+++ tests/test_plugin_manager.py
@@ -1,9 +1,88 @@
-import unittest
+from tests.testcase import HookeeTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 
 
-class TestPluginManager(unittest.TestCase):
-    pass
+class TestPluginManager(HookeeTestCase):
+    def test_validate_plugin(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_validate_plugin_not_conform_to_spec(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_validate_plugin_no_plugin_type(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_validate_plugin_no_run(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_validate_plugin_wrong_args(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_validate_plugin_no_blueprint(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_load_plugin_not_found(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_load_plugins(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_load_plugins_missing_required_plugin(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN