+++ tests/test_util.py
@@ -2,8 +2,33 @@ import unittest
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 
 
 class TestUtil(unittest.TestCase):
-    pass
+    def test_is_python_3(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def test_get_functions(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN
+
+    def get_args(self):
+        # TODO implement
+        pass
+        # GIVEN
+
+        # WHEN
+
+        # THEN