+++ hookee/conf.py
@@ -4,7 +4,7 @@ import confuse
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.11"
+__version__ = "0.0.12"
 
 template = {
     "auth_token": confuse.String(default=None),