+++ tests/managedtestcase.py
@@ -13,7 +13,7 @@ else:
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.11"
+__version__ = "0.0.12"
 
 
 class ManagedTestCase(HookeeTestCase):