+++ tests/test_util.py
@@ -6,7 +6,7 @@ from tests.testcase import HookeeTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "0.0.11"
+__version__ = "0.0.12"
 
 
 class TestUtil(HookeeTestCase):