+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/alexdlaird/hookee/compare/0.0.8...HEAD)
+### Changed
+- Replaced `response_json` and `response_xml` plugins with `response_echo` plugin.
+- Renamed `--list-plugins` to `--available-plugins`.
+
+### Fixed
+- Stability improvements.
 
 ## [0.0.8](https://github.com/alexdlaird/hookee/compare/0.0.7...0.0.8) - 2019-09-05
 ### Added